Set up state for the inputs of the first section
Information added by the user is stored in a object and this object
containing the information to be displayed in the CV preview will have
to be a variable shared between both the FullForm component and Preview
Component. Hence, the state is lifted to their parent component (App) so
when the user clicks submit, it triggers the set state action, updating
the state variable for the Preview component to display the new info.