Fixed bug of undefined value at the initial input
The previous code add the entry of the initial input from the user
when the page is loaded. The input element and its value was predefined
before the general info state was set. Hence when the user clicks the
submit button, the value it reads off was the initial value of input
when the DOM is loaded which was undefined.

The bug is fixed by updating the general info state whenever the input
element is changed. Therefore once the user finishes with inputting info
for a section and is about to click submit, the info will be updated
with the latest user input and stored as a state variable.

Another state is created to store the multiple entries of the info
submitted by the user. So whenever the user submits an entry, this entry
will be added to this "storage" state variable for the preview page to
render.