Added feature for edited entry to be resubmitted
Two states were created to track the whether the section being edited
and if it is, what is the id of that entry being edited.

Both state variables will be set when the user press the edit button,
so when the user decides to submit the entry again, there will be logic
deciding whether the entry is to be added as a new entry, or updated as
an existing entry.